Implemented a new method to do the Http Connection, and removed the files that did this work.
